refactor(e2e): split cancel-operation.spec.ts into focused test files
Extract accessibility tests into separate file for better organization:
- cancel-operation.spec.ts: Basic cancel operations (4 tests)
- accessibility.spec.ts: WAI-ARIA Dialog Pattern tests (3 tests)
  - Focus trap (Tab/Shift+Tab)
  - Background scroll disable
